More Refactoring #3
Apparently olc::v2d_generic::norm() does not compute the normal of a vector, rather it calculates the unit vector. Adjusted code accordingly. Additionally, slight optimization to rendering entities. They should only be rendered if they are actually on the screen. This saves computations when calculating unnecessary collisions and spending resources rendering entities that aren't even on the screen.