Reading from JSON
Thanks to the standalone JSON parser by nlohmann, we can now read in JSON files. This will help with constructing levels in the future. It is much more intuitive to edit a JSON with default parameters for each level to determine how many enemies, NPCs, and where to place everything. The map and player sprites are now loaded in using the `loadLevel(int level)` method. This is making huge progress towards #8.